+++ src/dftbp/dftbplus/initprogram.F90
@@ -66,6 +66,7 @@ module dftbp_dftbplus_initprogram
   use dftbp_dftb_thirdorder, only : TThirdOrderInp, TThirdOrder, ThirdOrder_init
   use dftbp_dftb_uniquehubbard, only : TUniqueHubbard, TUniqueHubbard_init
   use dftbp_dftb_elecconstraints, only : TElecConstraint, TElecConstraint_init, TElecConstraintInput
+  use dftbp_dftbplus_apicallback, only : TAPICallback
   use dftbp_dftbplus_elstattypes, only : elstatTypes
   use dftbp_dftbplus_forcetypes, only : forceTypes
   use dftbp_dftbplus_inputdata, only : TParallelOpts, TInputData, THybridXcInp, TControl, TBlacsOpts
@@ -1161,6 +1162,10 @@ module dftbp_dftbplus_initprogram
     !> of the atoms via the API is forbidden.
     logical :: atomOrderMatters = .false.
 
+    !> This object encapsulates subroutines and variables that are used for registering and
+    !! invocation of the density, overlap, and hamiltonian matrices exporting callbacks.
+    type(TAPICallback), allocatable :: apiCallBack
+
   #:if WITH_SCALAPACK
 
     !> Should the dense matrices be re-ordered for sparse operations
@@ -1375,6 +1380,14 @@ contains
       this%nSpin = 2
     end if
     this%nIndepSpin = this%nSpin
+
+  #:if WITH_API
+    if (input%ctrl%isASICallbackEnabled) then
+      allocate(this%apiCallBack)
+      ! As ASI doesn't set this, manually get from input whether H/S are expected to be modified:
+      this%apiCallBack%isAsiChangingTheModel = input%ctrl%isAsiChangingTheModel
+    end if
+  #:endif
 
     this%tSpinSharedEf = input%ctrl%tSpinSharedEf
     this%tSpinOrbit = input%ctrl%tSpinOrbit
@@ -1663,7 +1676,7 @@ contains
         & this%tFixEf, this%tSetFillingTemp, this%tFillKSep)
 
     call ensureSolverCompatibility(input%ctrl%solver%iSolver, this%kPoint, input%ctrl%parallelOpts,&
-        & this%nIndepSpin, this%tempElec)
+        & this%nIndepSpin, this%tempElec, input%ctrl%isASICallbackEnabled)
     call getBufferedCholesky_(this%tRealHS, this%parallelKS%nLocalKS, nBufferedCholesky)
     call TElectronicSolver_init(this%electronicSolver, input%ctrl%solver%iSolver, nBufferedCholesky)
 
@@ -3858,6 +3871,54 @@ contains
       end do
     end if
 
+    if (allocated(this%apiCallBack)) then
+      if (this%apiCallBack%isAsiChangingTheModel) then
+        if (allocated(this%scc)) then
+          ! as this needs assurances that the DM is being read by the external code, leading to
+          ! changes in the hamiltonian, otherwise SCC never converges.
+          call error("ASI callback with model modification enabled does not support self-consistent&
+              & calculations at present")
+        end if
+        if (this%tForces) then
+          ! Since if H and/or S is modified, the derivatives are not available via ASI at the
+          ! moment.
+          call error("ASI callback with model modification enabled does not support forces at&
+              & present")
+        end if
+        if (this%tMulliken) then
+          call error("ASI callback with model modification enabled does not support population&
+              & analysis at present")
+        end if
+      end if
+      if (allocated(this%reks)) then
+        call error("ASI callback does not support REKS")
+      end if
+      if (allocated(this%dftbU))  then
+        call error("ASI callback does not support +U at present")
+      end if
+      if (allocated(this%onSiteElements)) then
+        call error("ASI callback does currently support onsite corrections")
+      end if
+      if (this%isHybridXc) then
+        call error("ASI callback does currently support hybrid functionals")
+      end if
+      if (this%isElecDyn) then
+        call error("ASI callback does not currently support electron dynamics")
+      end if
+      if (this%tNegf) then
+        call error("ASI callback does not support transport")
+      end if
+      if (this%isLinResp) then
+        call error("ASI callback does not support linear response")
+      end if
+      if (allocated(this%ppRPA)) then
+        call error("ASI callback does not support ppRPA")
+      end if
+      if (this%doPerturbation) then
+        call error("ASI callback does not support perturbation calculations at present")
+      end if
+    end if
+
     if (this%deltaDftb%isNonAufbau) then
       if (this%nSpin /= 2) then
         call error("Internal error, Delta DFTB requires two spin channels")
@@ -5574,7 +5635,8 @@ contains
 
 
   !> Check for compatibility between requested electronic solver and features of the calculation
-  subroutine ensureSolverCompatibility(iSolver, kPoints, parallelOpts, nIndepSpin, tempElec)
+  subroutine ensureSolverCompatibility(iSolver, kPoints, parallelOpts, nIndepSpin, tempElec,&
+      & isCallBackApiEnabled)
 
     !> Solver number (see dftbp_elecsolvers_elecsolvertypes)
     integer, intent(in) :: iSolver
@@ -5590,6 +5652,9 @@ contains
 
     !> Temperature of the electrons
     real(dp), intent(in) :: tempElec
+
+    !> Is the API callback for H/S/density matrix enabled?
+    logical, intent(in) :: isCallBackApiEnabled
 
     logical :: tElsiSolver
     integer :: nKPoint
@@ -5617,6 +5682,13 @@ contains
 
     if (iSolver == electronicSolverTypes%pexsi .and. tempElec < epsilon(0.0)) then
       call error("This solver requires a finite electron broadening")
+    end if
+
+    if (isCallBackApiEnabled .and. .not. any(iSolver == [electronicSolverTypes%qr,&
+        & electronicSolverTypes%divideandconquer, electronicSolverTypes%relativelyrobust,&
+        & electronicSolverTypes%elpa, electronicSolverTypes%magma_gvd])) then
+      ! there are not dense matrices for ASI to work with
+      call error("ASI interface incompatible with the current choice of electronic solver")
     end if
 
   end subroutine ensureSolverCompatibility